dijkstra json output

+++ transport_layer/src/heuristic_selector.cpp
@@ -16,7 +16,7 @@ heuristic_selector::heuristic_selector(float a, float b, float p, long seed, Dij
 
 t_edge* heuristic_selector::selected_edge(list<t_edge*> edges, int current_id, int dest_id, int time) {
     double total = 0, total_prob = 0, best_wait = 0;
-    float max_pheromone = 1.0f;
+    double max_pheromone = 1.0;
     
     // get maximum pheromone of all the edges available including "wait" pheromone
     for (list<t_edge*>::iterator it = edges.begin(); it != edges.end(); ++it) {
@@ -76,7 +76,7 @@ t_edge* heuristic_selector::selected_edge(list<t_edge*> edges, int current_id, i
     return nullptr;
 }
 
-double heuristic_selector::calculate_heuristic(int node_id, int dest_id, int e_dist, double ph, float max_pheromone) {
+double heuristic_selector::calculate_heuristic(int node_id, int dest_id, int e_dist, double ph, double max_pheromone) {
     float d = distance(node_id, dest_id, e_dist);
     
     if (ph == 0.0f) {
@@ -90,6 +90,6 @@ float heuristic_selector::distance(int node_id, int dest_id, int e_dist) {
     int distance = e_dist + d_map->get_edge_weight(node_id, dest_id);
     
     float t_d = distance / d_map->get_max_dj_distance();
-    
+
     return (1.0f - ((t_d >= 1.0f) ? distance / (e_dist + d_map->get_max_dj_distance() + 1) : t_d)) * 1000;
 }
